feat:fix

+++ src/main/java/com/scms/scms_be/model/entity/Manufacturing/BOMDetail.java
@@ -31,7 +31,7 @@ public class BOMDetail {
     @JoinColumn(name = "item_id", nullable = false)
     private Item item;
 
-    private Long quantity;
+    private Double quantity;
     private String note;
     
 }
